@@ -12,7 +12,8 @@
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@mui/material": "^5.4.1",
-    "next": "12.0.10",
+    "@mui/styles": "^5.4.2",
+    "next": "^12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "sass": "^1.49.7"
