@@ -1,30 +1,30 @@
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+import { createStyles, makeStyles, DefaultTheme } from "@mui/styles";
 
-export const GameOfLifeStyles = makeStyles((theme: Theme) =>
+export const GameOfLifeStyles = makeStyles((theme: DefaultTheme) =>
   createStyles({
     marginTop4vhSpacing: {
-        marginTop: '4vh'
+      marginTop: "4vh",
     },
     formContainer: {
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-        marginTop: '4vh'
+      display: "flex",
+      justifyContent: "center",
+      alignItems: "center",
+      marginTop: "4vh",
     },
     formControl: {
       minWidth: 140,
-      marginRight: '4vh'
+      marginRight: "4vh",
     },
     center: {
-        textAlign: 'center'
+      textAlign: "center",
     },
     hide: {
-      display: 'none'
+      display: "none",
     },
     gridContainer: {
-        marginTop: '8vh',
-        display: 'flex',
-        justifyContent: 'center'
-    }
+      marginTop: "8vh",
+      display: "flex",
+      justifyContent: "center",
+    },
   })
 );
