@@ -1,13 +1,13 @@
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+import { createStyles, makeStyles, DefaultTheme } from "@mui/styles";
 
-export const TopNavStyles = makeStyles((theme: Theme) =>
+export const TopNavStyles = makeStyles((theme: DefaultTheme) =>
   createStyles({
     root: {
       flexGrow: 1,
     },
     title: {
       flexGrow: 1,
-      textAlign: 'center'
+      textAlign: "center",
     },
-  }),
+  })
 );
